Move makeDraggable to a separate file
