+++ src/components/dialog/DialogBase.js
@@ -11,6 +11,7 @@ import Footer from './Footer';
 import theme from './theme.css';
 import uiUtilities from '@teamleader/ui-utilities';
 import useFocusTrap from '../../utils/useFocusTrap';
+import { makeDraggable } from '../../utils/makeDraggable/makeDraggable';
 
 export const DialogBase = ({
   active,
@@ -74,43 +75,6 @@ export const DialogBase = ({
   );
 
   return createPortal(dialog, document.body);
-};
-
-const makeDraggable = (dragTargetRef, dragHandleRef) => {
-  if (!dragTargetRef?.current || !dragHandleRef?.current) {
-    return;
-  }
-
-  const ele = dragTargetRef.current;
-  const handle = dragHandleRef.current;
-  let x = 0;
-  let y = 0;
-
-  const mouseDownHandler = function (e) {
-    x = e.clientX;
-    y = e.clientY;
-
-    document.addEventListener('mousemove', mouseMoveHandler);
-    document.addEventListener('mouseup', mouseUpHandler);
-  };
-
-  const mouseMoveHandler = function (e) {
-    const dx = e.clientX - x;
-    const dy = e.clientY - y;
-
-    ele.style.top = `${ele.offsetTop + dy}px`;
-    ele.style.left = `${ele.offsetLeft + dx}px`;
-
-    x = e.clientX;
-    y = e.clientY;
-  };
-
-  const mouseUpHandler = function () {
-    document.removeEventListener('mousemove', mouseMoveHandler);
-    document.removeEventListener('mouseup', mouseUpHandler);
-  };
-
-  handle.addEventListener('mousedown', mouseDownHandler);
 };
 
 DialogBase.propTypes = {
